roll sections given their own components
